user permission service WIP

+++ src/main/java/com/cub/project/api/controllers/GroupController.java
@@ -5,6 +5,7 @@ import com.cub.project.domain.models.Group;
 import com.cub.project.domain.models.Post;
 import com.cub.project.domain.models.User;
 import com.cub.project.service.GroupService;
+import com.cub.project.service.UserPermissionService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -20,51 +21,63 @@ import java.util.Collection;
 @RequiredArgsConstructor
 public class GroupController {
     private final GroupService groupService;
+    private final UserPermissionService permissionService;
 
-    @PreAuthorize("")
     @GetMapping("{groupId}")
     public ResponseEntity<Group> getGroupData(@PathVariable long groupId, @AuthenticationPrincipal User auth) {
-        return new ResponseEntity<>(groupService.getGroupById(groupId), new HttpHeaders(), HttpStatus.OK);
+        if (permissionService.isMember(groupId, auth.getEmail())) {
+            return new ResponseEntity<>(groupService.getGroupById(groupId), new HttpHeaders(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("")
     @PostMapping("new")
     public ResponseEntity<?> createGroup(@RequestBody GroupDto group, @AuthenticationPrincipal User auth) {
         groupService.createGroup(group);
         return ResponseEntity.ok().build();
     }
 
-    @PreAuthorize("")
     @GetMapping("{groupId}/posts")
     public ResponseEntity<Collection<Post>> getPosts(@PathVariable long groupId, @AuthenticationPrincipal User auth) {
-        return new ResponseEntity<>(groupService.getGroupById(groupId).getPosts(), new HttpHeaders(), HttpStatus.OK);
+        if (permissionService.isMember(groupId, auth.getEmail())) {
+            return new ResponseEntity<>(groupService.getGroupById(groupId).getPosts(), new HttpHeaders(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("")
     @PatchMapping("{groupId}/remove/participant/{participantId}")
     public ResponseEntity<?> removeParticipant(@PathVariable long groupId, @PathVariable long userId, @AuthenticationPrincipal User auth) {
-        groupService.removeParticipant(groupId, userId);
-        return ResponseEntity.ok().build();
+        if (permissionService.isAdmin(groupId, auth.getEmail())) {
+            groupService.removeParticipant(groupId, userId);
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("")
     @PatchMapping("{groupId}/remove/post/{postId}")
     public ResponseEntity<?> removePost(@PathVariable long groupId, @PathVariable long postId, @AuthenticationPrincipal User auth) {
-        groupService.removePost(groupId, postId);
-        return ResponseEntity.ok().build();
+        if (permissionService.isAdmin(groupId, auth.getEmail())) {
+            groupService.removePost(groupId, postId);
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("")
     @PatchMapping("{groupId}/edit")
     public ResponseEntity<?> editGroup(@PathVariable long groupId, @RequestBody GroupDto group, @AuthenticationPrincipal User auth) {
-        groupService.updateGroup(group);
-        return ResponseEntity.ok().build();
+        if (permissionService.isAdmin(groupId, auth.getEmail())) {
+            groupService.updateGroup(group);
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("")
     @DeleteMapping("{groupId}/delete")
     public ResponseEntity<?> deleteGroup(@PathVariable long groupId, @AuthenticationPrincipal User auth) {
-        groupService.deleteGroup(groupId);
-        return ResponseEntity.ok().build();
+        if (permissionService.isAdmin(groupId, auth.getEmail())) {
+            groupService.deleteGroup(groupId);
+            return ResponseEntity.ok().build();
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 }
